ml/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.2
lightgbm==4.5.0
torch==2.5.0
shap==0.46.0
langchain==0.3.0
langchain-anthropic==0.3.0
chromadb==0.5.0
sentence-transformers==3.3
scipy==1.14.1
statsmodels==0.14.4
opencv-python==4.10.0
mlflow==2.17.0
optuna==4.0.0
joblib==1.4.2
python-dotenv==1.0.1
requests==2.32.3
matplotlib==3.9.2
seaborn==0.13.2
ipykernel==6.29.5
